sets up roommates page and one view page
