+++ app/Components/Cards/RoommateCard.tsx
@@ -1,40 +1,30 @@
-
-import Link from "next/link";
-
-export default function UserCard({
-    firstN, 
-    lastN,
-    gender,
-    age,
+export default function RoommateCard({
     id,
-    email,
-
+    bio,
+    budget,
+    preferredGender,
+    smoke,
 }:{
-    firstN: string;
-    lastN: string;
-    gender: 'male' | 'female' | 'other';
-    age: number;
-    id: number;
-    email: string;
-}){
+    id:number;
+    bio: string;
+    budget: number;
+    preferredGender: 'male' | 'female' | 'other' | 'no preference';
+    smoke: 'no' | 'yes';
+}) {
+    
     return (
         <>
-           <div className="card w-15 m-3 bg-indigo-500 shadow-xl">
+            <div className="card w-15 m-3 bg-indigo-500 shadow-xl">
                 {/* replace img element with built-in nextjs Image component */}
                 <figure></figure>
                 <div className="card-body">
-                    <h2 className="card-title">{firstN} {lastN}</h2>
-                    <p>Gender: {gender}</p>
-                    <p>Age: {age}</p>
-                    <p>email: {email}</p>
-
+                    <h2 className="card-title">Budget: {budget}</h2>
+                    <p>Bio: {bio}</p>
+                    <p>Preferred Gender: {preferredGender}</p>
+                    <p>Smoking: {smoke}</p>
                     <p>{id}</p>
-
-                    <div className="card-actions justify-end">
-                    <Link href={`/roommates/${id}`}><button className="btn btn-primary">View</button></Link>
-                    </div>
                 </div>
-            </div> 
+            </div>
         </>
     )
 }
