+++ app/roommates/[id]/page.tsx
@@ -1,9 +1,12 @@
-import UserCard from "@/app/Components/Cards/RoommateCard";
+import UserCard from "@/app/Components/Cards/UserCard";
+import RoommateCard from "@/app/Components/Cards/RoommateCard";
+
 import { getOneRoommate } from "@/app/lib/data";
 
 export default async function OneRoommate({params}: {params: { id: number}}) {
     const id = params.id
     const user = await getOneRoommate(id);
+    
 
     return (
         <>
@@ -13,6 +16,7 @@ export default async function OneRoommate({params}: {params: { id: number}}) {
                     gender={user.gender} age={user.age}
                     id={user.id} email={user.email} />}
             </div>
+           { user && <RoommateCard id={user.formid} budget={user.budget} bio={user.bio} preferredGender={user.preferredgender} smoke={user.smokes} />}
         </>
     )
