+++ app/roommates/page.tsx
@@ -1,5 +1,5 @@
 import { fetchRoommates } from "../lib/data";
-import UserCard from "../Components/Cards/RoommateCard";
+import UserCard from "../Components/Cards/UserCard";
 
 export default async function RoommatesPage() {
     const users = await fetchRoommates();
